requirements.txt
# Base dependencies for OpenSSME
# Add more as needed per phase
numpy
pandas
pyyaml
mkdocs
mkdocs-material
pypdf
python-dotenv
google-generativeai
pygments
ipykernel
